MAINT: Move SMDA health check info to separate component
